api_assignment/api_assignment/requirements.txt
Pillow == 8.4.0
django-thumbnails == 0.7.0
Django==3.2.18
djangorestframework == 3.14.0
